requirements.txt
discord.py[voice]
discord-ext-voice-recv
faster-whisper
ollama
chromadb
python-dotenv
websockets
numpy
scipy
fastapi
uvicorn
requests
